+++ projects/swimlane/ngx-ui/src/lib/components/dropzone/dropzone.module.ts
@@ -4,10 +4,11 @@ import { CommonModule } from '@angular/common';
 import { DropzoneComponent } from './dropzone.component';
 import { ButtonModule } from '../button/button.module';
 import { IconModule } from '../icon/icon.module';
+import { TooltipModule } from '../tooltip/tooltip.module';
 
 @NgModule({
   declarations: [DropzoneComponent],
   exports: [DropzoneComponent],
-  imports: [CommonModule, ButtonModule, IconModule]
+  imports: [CommonModule, ButtonModule, IconModule, TooltipModule]
 })
 export class DropzoneModule {}